Updated panels to handle appended side panels
The previous set up only allowed us to append side panels to the LayersPanel index.tsx (if we want to use the blur effect). Using the same backdrop blur on a child when the parent has the blur as well wouldn't work.

The new setup lets us append side panels to the specific tab index.tsx and maintain the same background blur. Little confusing - can explain this more if needed.