App Panel, Installed Apps, App Details (#1608)
* Apps panel base created

* Detail Panel base added with Open State

* Updated panels to handle appended side panels

The previous set up only allowed us to append side panels to the LayersPanel index.tsx (if we want to use the blur effect). Using the same backdrop blur on a child when the parent has the blur as well wouldn't work.

The new setup lets us append side panels to the specific tab index.tsx and maintain the same background blur. Little confusing - can explain this more if needed.

* Updated minor styling

* Updated styling & Organized components

* Updated tools accordion styles

* Updated Styling

* Styling updates

* Updated spacing and fonts

* Fixed panel width problems

* Consolidated the AppCard & FeaturedAppCard

* Installed tab initial base - needs styling

* AppCard Hover/Active effects

* Installed apps - base card styled

* Updated InstalledCard error states

* Updated settings style

* Updated InstallCard Error

* Updated status indicators

* Detail Panel installed error bar

* Authenticate by link added

* Updated button styles

* Added various states

* Updated hidden toolbar

* Add missing translations