Fix top bar scale so external link does not hide (#2217)
* Fix top bar scale so external link does not hide

Fix the scaling on the top bar so that the external link button does not
get hidden when you zoom out too much.

* Make solution more simple

* change scale origin for better padding

* clean up

* Add cutoff for top bar icons

I also added an empty top bar that shows so that the user can drag the
window around

* Use inline condition instead of early return

I wasn't sure what would be better, but I'm leaning towards this because
it means you don't have to duplicate the code for the top bar when
returning an empty top bar early.

* Update changes to preload